+++ packages/design-system-react-native/src/components/SectionHeader/index.ts
@@ -1,0 +1,3 @@
+export type { SectionHeaderPropsShared } from '@metamask/design-system-shared';
+export { SectionHeader } from './SectionHeader';
+export type { SectionHeaderProps } from './SectionHeader.types';