+++ packages/design-system-react-native/src/components/index.ts
@@ -273,3 +273,9 @@ export type { BannerAlertProps } from './BannerAlert';
 
 export { Tag, TagSeverity } from './Tag';
 export type { TagProps } from './Tag';
+
+export { SectionHeader } from './SectionHeader';
+export type {
+  SectionHeaderProps,
+  SectionHeaderPropsShared,
+} from './SectionHeader';